Refactor old integration tests into separate mods
