+++ tests/integration/runners/vault/conftest.py
@@ -1,13 +1,34 @@
+import shutil
+from pathlib import Path
+
 import pytest
 
-from tests.support.vault import vault_delete_secret
-from tests.support.vault import vault_write_secret
+
+@pytest.fixture(scope="module")
+def _minion_conn_cachedir(minion):
+    ret = minion.salt_call_cli().run("config.get", "cachedir")
+    assert ret.returncode == 0
+    assert ret.data
+    return Path(ret.data) / "vault" / "connection"
 
 
-@pytest.fixture(scope="class")
-def vault_testing_values(container):  # pylint: disable=unused-argument
-    vault_write_secret("secret/path/foo", success="yeehaaw")
-    try:
-        yield
-    finally:
-        vault_delete_secret("secret/path/foo")
+@pytest.fixture
+def minion_conn_cachedir(_minion_conn_cachedir):
+    if not _minion_conn_cachedir.exists():
+        _minion_conn_cachedir.mkdir(parents=True)
+    return _minion_conn_cachedir
+
+
+@pytest.fixture
+def conn_cache_absent(minion_conn_cachedir):
+    if minion_conn_cachedir.exists():
+        shutil.rmtree(minion_conn_cachedir)
+        assert not minion_conn_cachedir.exists()
+    yield
+
+
+@pytest.fixture(scope="module")
+def vault_secrets_defaults():
+    return {
+        "secret/path/foo": {"success": "yeehaaw"},
+    }
