+++ tests/integration/runners/vault/test_master_cluster.py
@@ -14,7 +14,7 @@ sver = int(salt.version.__version__.split(".")[0])
 pytestmark = [
     pytest.mark.skip_if_binaries_missing("vault"),
     pytest.mark.skipif(sver < 3007, reason="Master cluster requires Salt 3007+"),
-    pytest.mark.usefixtures("container", "vault_testing_values"),
+    pytest.mark.usefixtures("container", "vault_secrets"),
 ]
