+++ tests/integration/utils/factory/test_sdb_in_config.py
@@ -1,5 +1,6 @@
 import pytest
 
+from tests.conftest import CONTAINER_TARGETS
 from tests.support.helpers import PatchedEnviron
 from tests.support.vault import vault_create_secret_id
 from tests.support.vault import vault_delete_approle
@@ -10,6 +11,9 @@ from tests.support.vault import vault_write_approle
 
 pytestmark = [
     pytest.mark.skip_if_binaries_missing("vault"),
+    pytest.mark.parametrize(
+        "container", (CONTAINER_TARGETS[0],), indirect=True
+    ),  # We only want to check the internal logic, not the API access
 ]
